onlinejudge/requirements.txt
Django==5.0.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
djoser==2.3.1
gunicorn
markdown
dotenv
httpx